// lib/api/generated/models/CreateApplicationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateApplicationDto = {
    coverLetter: string;
    resumeUrl?: string;
    portfolioUrl?: string;
    phone?: string;
    linkedinUrl?: string;
    githubUrl?: string;
};